saving progress
grug responds to "hey grug" in voice chats and gives a ping sound when he's been triggered
