+++ .gitattributes
@@ -1,0 +1,1 @@
+*.wav filter=lfs diff=lfs merge=lfs -text
