+++ pyproject.toml
@@ -27,6 +27,8 @@ dependencies = [
     "discord-ext-voice-recv>=0.4.2a145",
     "speechrecognition[openai]>=3.14.0",
     "tembo-pgmq-python[async]>=0.9.0",
+    "gradio-tools>=0.0.9",
+    "rapidfuzz>=3.12.1",
 ]
 
 [dependency-groups]
